midnight-cosmetics-agent/requirements.txt
claude-agent-sdk>=0.1.50
anthropic>=0.52.0
openai>=1.82.0
streamlit>=1.45.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
